refactor: reduce duplicated code about identity
